Split out actions into their composite classes
