+++ lib/cucumber/core/test/actions.rb
@@ -5,3 +5,5 @@ require 'cucumber/core/test/result'
 require 'cucumber/core/test/timer'
 
 require 'cucumber/core/test/actions/action'
+require 'cucumber/core/test/actions/undefined_action'
+require 'cucumber/core/test/actions/unskippable_action'
